+++ app/Http/Controllers/SertifikatController.php
@@ -7,37 +7,34 @@ use Illuminate\Http\Request;
 
 class SertifikatController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
     public function index(Request $request)
     {
-        $search = $request->input('search'); // untuk search berdasarkan nama sertifikat
+        if (!auth()->user()->can('unduh_sertifikat.list')) return abort(403);
 
-        $sertifikat = Sertifikat::query();
+        $search = $request->input('search');
+
+        $data = Sertifikat::query();
 
         if ($search) {
-            $sertifikat->where('nama_sertifikat', 'like', "%{$search}%");
+            $data->where('nama_sertifikat', 'like', "%{$search}%");
         }
 
-        $sertifikat = $sertifikat->paginate(10);
+        $data = $data->latest()->paginate(10);
 
-        return view('sertifikat.index', compact('sertifikat', 'search'));
+        return view('sertifikat.index', compact('data'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     */
     public function create()
     {
+        if (!auth()->user()->can('unduh_sertifikat.buat')) return abort(403);
+
         return view('sertifikat.create');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(Request $request)
     {
+        if (!auth()->user()->can('unduh_sertifikat.buat')) return abort(403);
+
         $request->validate([
             'file_pdf' => 'required|mimes:pdf|max:2048',
         ]);
@@ -55,29 +52,21 @@ class SertifikatController extends Controller
         return redirect()->route('sertifikat.index')->with('status', 'Sertifikat berhasil diupload');
     }
 
-    /**
-     * Display the specified resource.
-     */
-    public function show(Sertifikat $sertifikat)
-    {
-        return view('sertifikat.show', compact('sertifikat'));
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     */
     public function edit(Sertifikat $unduh_sertifikat)
     {
-        return view('sertifikat.edit', compact('sertifikat'));
+        if (!auth()->user()->can('unduh_sertifikat.edit')) return abort(403);
+
+        return view('sertifikat.edit', [
+            'sertifikat' => $unduh_sertifikat
+        ]);
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
     public function update(Request $request, Sertifikat $unduh_sertifikat)
     {
+        if (!auth()->user()->can('unduh_sertifikat.edit')) return abort(403);
+
         $request->validate([
-            'file_pdf' => 'mimes:pdf|max:2048',
+            'file_pdf' => 'required|mimes:pdf|max:2048',
         ]);
 
         if ($request->hasFile('file_pdf')) {
@@ -101,11 +90,10 @@ class SertifikatController extends Controller
         return redirect()->route('sertifikat.index')->with('status', 'Sertifikat berhasil diupdate');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
     public function destroy(Sertifikat $unduh_sertifikat)
     {
+        if (!auth()->user()->can('unduh_sertifikat.hapus')) return abort(403);
+
         if ($unduh_sertifikat->nama_file && file_exists(storage_path('app/public/sertifikat/' . $unduh_sertifikat->nama_file))) {
             unlink(storage_path('app/public/sertifikat/' . $unduh_sertifikat->nama_file));
         }
@@ -115,11 +103,10 @@ class SertifikatController extends Controller
         return redirect()->route('sertifikat.index')->with('status', 'Sertifikat berhasil dihapus');
     }
 
-    /**
-     * Download the specified sertifikat.
-     */
     public function download(Sertifikat $sertifikat)
     {
+        if (!auth()->user()->can('unduh_sertifikat.download')) return abort(403);
+
         $path = storage_path('app/public/sertifikat/' . $sertifikat->nama_file);
 
         if (file_exists($path)) {
