+++ app/Http/Controllers/SupervisiKepruController.php
@@ -19,6 +19,8 @@ class SupervisiKepruController extends Controller
 
     public function index(Request $request)
     {
+        if (!auth()->user()->can('supervisi_kepru.list')) return abort(403);
+
         $data = $this->service->getAll($request);
 
         return view('spv_kepru.index', compact('data'));
@@ -26,6 +28,8 @@ class SupervisiKepruController extends Controller
 
     public function create()
     {
+        if (!auth()->user()->can('supervisi_kepru.buat')) return abort(403);
+
         return view('spv_kepru.create', [
             'route' => route('spv_kepru.store'),
             'method' => 'POST'
@@ -34,6 +38,8 @@ class SupervisiKepruController extends Controller
 
     public function store(SpvKepruRequest $request)
     {
+        if (!auth()->user()->can('supervisi_kepru.buat')) return abort(403);
+
         $request->validated();
 
         $this->service->store($request);
@@ -48,6 +54,8 @@ class SupervisiKepruController extends Controller
 
     public function edit(SpvKepru $spv_kepru)
     {
+        if (!auth()->user()->can('supervisi_kepru.edit')) return abort(403);
+
         return view('spv_kepru.edit', [
             'spv_kepru' => $spv_kepru,
             'route' => route('spv_kepru.update', $spv_kepru->id),
@@ -57,6 +65,8 @@ class SupervisiKepruController extends Controller
 
     public function update(SpvKepruRequest $request, SpvKepru $spv_kepru)
     {
+        if (!auth()->user()->can('supervisi_kepru.edit')) return abort(403);
+
         $request->validated();
 
         $this->service->update($spv_kepru, $request);
@@ -66,6 +76,8 @@ class SupervisiKepruController extends Controller
 
     public function destroy(SpvKepru $spv_kepru)
     {
+        if (!auth()->user()->can('supervisi_kepru.hapus')) return abort(403);
+
         $this->service->delete($spv_kepru);
         return redirect()->route('spv_kepru.index')->with('success', 'Supervisi berhasil dihapus.');
     }
